getHilfeSolution/modules/ingestintodb/requirements.txt
azure-iot-device~=2.7.0
datetime
backports.zoneinfo
mqtt
paho-mqtt
SQLAlchemy
psycopg2-binary